requirements.txt
asgiref==3.7.2
Django==4.2.9
django-bootstrap5==23.4
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
